version workflow and change to test

+++ mct_quantizers/keras/quantizers/activation_inferable_quantizers/activation_lut_pot_inferable_quantizer.py
@@ -127,7 +127,7 @@ if FOUND_TF:
                                                f'{inputs.dtype}'
 
             return lut_quantizer(inputs,
-                                 cluster_centers=self._cluster_centers_as_np,
+                                 cluster_centers=self._cluster_centers_as_np.astype(np.float32),
                                  signed=self.signed,
                                  # In activation per-channel quantization is not supported
                                  # thus we expect a single threshold value. Assertion is made in init.
